mydjango/requirements/dev.txt
-r test.txt

ipython

django-debug-toolbar
django-extensions

